pyproject.toml
[project]
name = "environmental-risk-metrics"
version = "0.1.14"
description = "Calculate environmental risk metrics for a given polygon"
readme = "README.md"
authors = [
    { name = "Thimm Zwiener", email = "zwiener@gmail.com" },
    { name = "Thimm Zwiener", email = "thimm@regenrate.com" }
]
requires-python = ">=3.10"
dependencies = [
    "geopandas>=1.0.1",
    "geopy>=2.4.1",
    "leafmap>=0.42.5",
    "localtileserver>=0.10.5",
    "odc-stac>=0.3.10",
    "pint>=0.24.4",
    "pint-pandas>=0.6.2",
    "planetary-computer>=1.0.0",
    "plotly>=5.24.1",
    "pyarrow>=18.1.0",
    "pydantic-pint>=0.1",
    "pygbif>=0.6.5",
    "pystac-client>=0.8.5",
    "rasterstats>=0.20.0",
    "rioxarray>=0.18.1",
    "scipy>=1.14.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = ["./src/"]
filterwarnings = [
    "ignore:Support for class-based `config` is deprecated:DeprecationWarning"
]

[dependency-groups]
dev = [
    "pytest-mock>=3.14.0",
    "pytest>=8.3.4",
    "jupyter>=1.1.1",
]
